PostsController update

+++ laravel-app/app/Http/Controllers/PostsController.php
@@ -87,7 +87,14 @@ class PostsController extends Controller
      */
     public function edit($id)
     {
+
         $post = Post::find($id);
+
+        //Check for correct user
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('posts')->with('error', 'You are not authorized.');
+        }
+
         return view('posts.edit')->with('post', $post);
     }
 
@@ -123,6 +130,12 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        //Check for correct user
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('posts')->with('error', 'You are not authorized.');
+        }
+
         $post->delete();
         return redirect('/posts')->with('success', 'Post Deleted');
     }
